Implement SuggestionsPopup in EditInputField
Create a mixin to easily apply SuggestionsPopup logic to EditInputField
and EditInputGroup without duplicating logic
