+++ src/components/EditInputField.vue
@@ -1,5 +1,5 @@
 <template>
-    <fieldset @keyup="keyPressed">
+    <fieldset>
         <label>
             {{ fieldLabel }}
         </label>
@@ -8,6 +8,9 @@
             ref="inputField"
             v-model="content"
             @input="handleInput"
+            @keydown="keyDown"
+            @keyup="handleSuggestionsPopupKeyUp"
+            @blur="handleSuggestionsPopupBlur"
         />
         <div v-else>
             <slot />
@@ -17,14 +20,30 @@
                 v-model="content"
                 :type="fieldType"
                 @input="handleInput"
+                @keydown="keyDown"
+                @keyup="handleSuggestionsPopupKeyUp"
+                @blur="handleSuggestionsPopupBlur"
             />
         </div>
+        <SuggestionsPopup
+            v-if="suggestionsData !== null"
+            ref="suggestionsPopup"
+            :offset="popupOffset"
+            :focus-index="suggestionsData.focusIndex"
+            :suggestion-options="filteredSelectionOptions"
+        />
     </fieldset>
 </template>
 
 <script>
+import SuggestionsPopup, { suggestionsPopupMixin } from "./SuggestionsPopup.vue"
+
 export default {
     name: "EditInputField",
+    components: {
+        SuggestionsPopup,
+    },
+    mixins: [suggestionsPopupMixin],
     props: {
         fieldLabel: {
             type: String,
@@ -65,55 +84,10 @@ export default {
         handleInput() {
             this.$emit("input", this.content)
         },
-        /**
-         * Catches # key down presses and opens recipe-references dialog
-         */
-        keyPressed(e) {
-            // Using keyup for trigger will prevent repeat triggering if key is held down
-            if (this.referencePopupEnabled && e.key === "#") {
-                e.preventDefault()
-                // Check if the letter before the hash
-                if (this.fieldType === "markdown") {
-                    // for reference: https://codemirror.net/doc/manual.html#api
-                    const cursorPos = JSON.parse(
-                        JSON.stringify(
-                            this.$refs.inputField.editor.getCursor("start")
-                        )
-                    )
-                    const prevChar = this.$refs.inputField.editor.getRange(
-                        { line: cursorPos.line, ch: cursorPos.ch - 2 },
-                        { line: cursorPos.line, ch: cursorPos.ch - 1 }
-                    )
-                    if (
-                        cursorPos.ch === 1 ||
-                        prevChar === " " ||
-                        prevChar === "\n" ||
-                        prevChar === "\r"
-                    ) {
-                        // beginning of line
-                        this.$parent.$emit("showRecipeReferencesPopup", {
-                            context: this,
-                        })
-                        this.lastCursorPosition = cursorPos
-                    }
-                } else {
-                    const cursorPos = this.$refs.inputField.selectionStart
-                    const content = this.$refs.inputField.value
-                    const prevChar =
-                        cursorPos > 1 ? content.charAt(cursorPos - 2) : ""
-                    if (
-                        cursorPos === 1 ||
-                        prevChar === " " ||
-                        prevChar === "\n" ||
-                        prevChar === "\r"
-                    ) {
-                        // Show dialog to select recipe
-                        this.$parent.$emit("showRecipeReferencesPopup", {
-                            context: this,
-                        })
-                        this.lastCursorPosition = cursorPos
-                    }
-                }
+        keyDown(e) {
+            // Redirect to suggestions handler if in suggestion mode
+            if (this.suggestionsData !== null) {
+                this.handleSuggestionsPopupKeyDown(e)
             }
         },
         pasteCanceled() {
