+++ src/components/EditInputGroup.vue
@@ -19,6 +19,7 @@
                     @keyup="keyUp"
                     @input="handleInput"
                     @paste="handlePaste"
+                    @blur="handleSuggestionsPopupBlur"
                 />
                 <textarea
                     v-else-if="fieldType === 'textarea'"
@@ -28,7 +29,7 @@
                     @keyup="keyUp"
                     @input="handleInput"
                     @paste="handlePaste"
-                    @blur="handleBlur"
+                    @blur="handleSuggestionsPopupBlur"
                 ></textarea>
                 <div class="controls">
                     <button
@@ -71,10 +72,7 @@
 </template>
 
 <script>
-import { position as caretPosition } from "caret-pos"
-import SuggestionsPopup, {
-    SUGGESTIONS_POPUP_WIDTH,
-} from "./SuggestionsPopup.vue"
+import SuggestionsPopup, { suggestionsPopupMixin } from "./SuggestionsPopup.vue"
 
 const linesMatchAtPosition = (lines, i) =>
     lines.every((line) => line[i] === lines[0][i])
@@ -96,13 +94,12 @@ const findCommonPrefix = (lines) => {
     }
 }
 
-const clamp = (val, min, max) => Math.min(max, Math.max(min, val))
-
 export default {
     name: "EditInputGroup",
     components: {
         SuggestionsPopup,
     },
+    mixins: [suggestionsPopupMixin],
     props: {
         value: {
             type: Array,
@@ -133,10 +130,6 @@ export default {
             type: Boolean,
             default: false,
         },
-        suggestionOptions: {
-            type: Array,
-            default: () => [],
-        },
     },
     data() {
         return {
@@ -145,34 +138,7 @@ export default {
             lastFocusedFieldIndex: null,
             lastCursorPosition: -1,
             ignoreNextKeyUp: false,
-            suggestionsData: null,
         }
-    },
-    computed: {
-        filteredSelectionOptions() {
-            const { searchText } = this.suggestionsData
-            return this.suggestionOptions.filter(
-                (option) =>
-                    searchText === "" ||
-                    option.title
-                        .toLowerCase()
-                        .includes(searchText.toLowerCase())
-            )
-        },
-
-        popupOffset() {
-            const { caretPos, field } = this.suggestionsData
-
-            return {
-                left: Math.min(
-                    field.offsetLeft + caretPos.left,
-                    field.offsetLeft +
-                        field.offsetWidth -
-                        SUGGESTIONS_POPUP_WIDTH
-                ),
-                top: field.offsetTop + caretPos.top + caretPos.height,
-            }
-        },
     },
     watch: {
         value: {
@@ -181,11 +147,6 @@ export default {
             },
             deep: true,
         },
-    },
-    mounted() {
-        this.$on("suggestions-selected", (opt) => {
-            this.handleSuggestionSelected(opt.recipe_id)
-        })
     },
     methods: {
         /* if index = -1, element is added at the end
@@ -315,57 +276,6 @@ export default {
             })
         },
         /**
-         * Cancel selection if input gets blurred
-         */
-        handleBlur(e) {
-            if (this.suggestionsData === null) return
-
-            // Do not cancel suggestions if the new focused element (e.relatedTarget)
-            // is a child of the suggestions popup
-            // That is the case when clicking an option in the suggestions popup,
-            // and cancelling too early prevents the option from being properly selected
-            if (this.$refs.suggestionsPopup[0].$el.contains(e.relatedTarget)) {
-                return
-            }
-            this.handleSuggestionsCancel()
-        },
-        /**
-         * Handle keydown events for suggestions popup
-         * The event will be sent here from the normal keydown handler
-         * if suggestionsData !== null
-         */
-        suggestionsHandleKeydown(e) {
-            // Handle switching the focused option with up/down keys
-            if (["ArrowUp", "ArrowDown"].includes(e.key)) {
-                e.preventDefault()
-
-                // Increment/decrement focuse index based on which key was pressed
-                // and constrain between 0 and length - 1
-                const focusIndex = clamp(
-                    this.suggestionsData.focusIndex +
-                        {
-                            ArrowUp: -1,
-                            ArrowDown: +1,
-                        }[e.key],
-                    0,
-                    this.filteredSelectionOptions.length - 1
-                )
-                this.suggestionsData = {
-                    ...this.suggestionsData,
-                    focusIndex,
-                }
-                return
-            }
-
-            // Handle selecting the current option when enter is pressed
-            if (e.key === "Enter") {
-                e.preventDefault()
-                const { focusIndex } = this.suggestionsData
-                const selection = this.filteredSelectionOptions[focusIndex]
-                this.handleSuggestionSelected(selection.recipe_id)
-            }
-        },
-        /**
          * Catches enter and key down presses and either adds a new row or focuses the one below
          */
         keyDown(e) {
@@ -373,6 +283,12 @@ export default {
             // ignore the following keyup event
             if (this.ignoreNextKeyUp) {
                 this.ignoreNextKeyUp = false
+                return
+            }
+
+            // Redirect to suggestions handler if in suggestion mode
+            if (this.suggestionsData !== null) {
+                this.handleSuggestionsPopupKeyDown(e)
                 return
             }
 
@@ -384,12 +300,6 @@ export default {
 
             // Repeat events should be ignored
             if (e.repeat) {
-                return
-            }
-
-            // Redirect to suggestions handler if in suggestion mode
-            if (this.suggestionsData !== null) {
-                this.suggestionsHandleKeydown(e)
                 return
             }
 
@@ -422,55 +332,6 @@ export default {
             }
         },
         /**
-         * Handle keyups events for suggestions popup
-         * The event will be sent here from the normal keydown handler
-         * if suggestionsData !== null
-         */
-        suggestionsHandleKeyUp(e, cursorPos) {
-            const caretPos = caretPosition(e.target, {
-                customPos: this.suggestionsData.hashPosition - 1,
-            })
-
-            // Only update the popover position if the line changes (caret pos y changes)
-            if (caretPos.top !== this.suggestionsData.caretPos.top) {
-                this.suggestionsData.caretPos = caretPos
-            }
-
-            // Cancel suggestion popup on whitespace or caret movement
-            if (
-                [
-                    " ",
-                    "\t",
-                    "#",
-                    "ArrowLeft",
-                    "ArrowRight",
-                    "Home",
-                    "End",
-                    "PageUp",
-                    "PageDown",
-                    "Escape",
-                ].includes(e.key)
-            ) {
-                this.handleSuggestionsCancel()
-                return
-            }
-
-            // Cancel suggestions popup if hash deleted
-            if (cursorPos < this.suggestionsData.hashPosition) {
-                this.handleSuggestionsCancel()
-                return
-            }
-
-            // Update the search text
-            // Slice the input from the position of the "#" to the caret position
-            const { hashPosition, field } = this.suggestionsData
-            const searchText = field.value.slice(hashPosition, cursorPos)
-            this.suggestionsData = {
-                ...this.suggestionsData,
-                searchText,
-            }
-        },
-        /**
          * Shows the recipe linking popup when # is pressed
          */
         keyUp(e) {
@@ -481,49 +342,14 @@ export default {
                 return
             }
 
-            // Get the index of the pressed list item
             const $li = e.currentTarget.closest("li")
             const $ul = $li.closest("ul")
             const $pressedLiIndex = Array.prototype.indexOf.call(
                 $ul.childNodes,
                 $li
             )
-
-            // Get the position of the cursor and the content of the input
-            const elm = this.$refs["list-field"][$pressedLiIndex]
-            const cursorPos = elm.selectionStart
-
-            if (this.suggestionsData !== null) {
-                this.suggestionsHandleKeyUp(e, cursorPos)
-                return
-            }
-
-            // Only do anything for enter or # keys
-            if (!(this.referencePopupEnabled && e.key === "#")) {
-                return
-            }
-
-            // Show the popup only if the # was inserted at the very
-            // beggining of the input or after any whitespace character
-            if (
-                !(cursorPos === 1 || /\s/.test(elm.value.charAt(cursorPos - 2)))
-            ) {
-                return
-            }
-
-            // Show dialog to select recipe
-            const caretPos = caretPosition(elm, { customPos: cursorPos - 1 })
-            this.suggestionsData = {
-                field: elm,
-                context: this,
-                searchText: "",
-                caretPos,
-                focusIndex: 0,
-                hashPosition: cursorPos,
-                fieldIndex: $pressedLiIndex,
-            }
             this.lastFocusedFieldIndex = $pressedLiIndex
-            this.lastCursorPosition = cursorPos
+            this.handleSuggestionsPopupKeyUp(e)
         },
         moveEntryDown(index) {
             if (index >= this.buffer.length - 1) {
@@ -586,9 +412,6 @@ export default {
         },
         handleSuggestionSelected(recipeId) {
             this.pasteString(`r/${recipeId} `)
-            this.suggestionsData = null
-        },
-        handleSuggestionsCancel() {
             this.suggestionsData = null
         },
     },
